To separate box score and game logs into custom files, create api_helper to keep methods for generate_headers and retry_with_backoff
